Add files via upload
This is a simple one-page website featuring the fictitious Omnicook brand. This is built and taught from a Udemy course [RWD and Development with HTML5 and CSS3](https://www.udemy.com/design-and-develop-a-killer-website-with-html5-and-css3/). 
Some things in this website has been changed by the user like 'Jquery, CSS' etc. An orignal version will be uploaded in due time.